Reorganise, add readme
