+++ src/dqmdisplay/apps/templates/event_display.html
