+++ src/dqmdisplay/apps/templates/event_display_plane.html
