+++ src/dqmdisplay/apps/templates/plot_navigator.html
